requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Deep Learning
tensorflow>=2.10.0
# tensorflow-cpu>=2.10.0  # For CPU-only systems

# Model Explainability
shap>=0.41.0

# Feature Store & MLOps
hopsworks>=3.0.0
joblib>=1.2.0

# API & Web
streamlit>=1.20.0
requests>=2.28.0
aiohttp>=3.8.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0

# Open-Meteo SDK (recommended for proper binary parsing)
openmeteo_requests>=1.0.0

# Data Visualization
plotly>=5.13.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.0

# Configuration & Logging
python-dotenv>=1.0.0
loguru>=0.6.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: Performance
numba>=0.57.0
bottleneck>=1.3.0

# Optional: Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
